+++ babel.config.js
@@ -1,6 +1,3 @@
-const inProduction = process.env.NODE_ENV === "production";
-const inStorybookBuild = process.env.STORYBOOK_BUILD === "true";
-
 module.exports = (api) => {
   const useCJSModules = api.env(["test", "cjs"]);
 
@@ -22,9 +19,6 @@ module.exports = (api) => {
       "@babel/preset-typescript",
     ],
     plugins: [
-      !inStorybookBuild &&
-        inProduction &&
-        "./plugins/generate-proptypes/index.mjs",
       "@babel/plugin-proposal-class-properties",
       "@babel/plugin-transform-classes",
       "babel-plugin-dev-expression",