+++ jest.config.ts
@@ -1,7 +1,9 @@
 import { Config } from "jest";
+import { resolve } from "path";
 
 import coverageThresholds from "./coverage-thresholds.json";
 
+const rootDir = resolve(__dirname, ".");
 const esmOnlyPackages = [
   "react-dnd",
   "core-dnd",
@@ -9,31 +11,10 @@ const esmOnlyPackages = [
   "dnd-core",
   "react-dnd-html5-backend",
 ];
-
 const isCI = process.env.CI === "true";
 
-const config: Config = {
-  notify: false,
-  setupFiles: ["raf/polyfill"],
-  testEnvironment: "jsdom",
-  setupFilesAfterEnv: [
-    "<rootDir>/src/__spec_helper__/__internal__/index.ts",
-    "jest-canvas-mock",
-  ],
-  testMatch: ["**/?(*.)+(spec|test).[jt]s?(x)"],
-  testPathIgnorePatterns: ["node_modules", "lib", "esm"],
-  moduleDirectories: ["src", "node_modules"],
-  collectCoverage: true,
-  coveragePathIgnorePatterns: [
-    "node_modules",
-    "src/__spec_helper__",
-    "src/locales",
-    "lib",
-    "esm",
-  ],
-  coverageReporters: ["text-summary", "html"],
-  coverageDirectory: "<rootDir>/coverage",
-  coverageThreshold: isCI ? undefined : coverageThresholds,
+const baseProjectConfig: Config = {
+  rootDir,
   moduleFileExtensions: ["ts", "tsx", "js", "jsx", "mjs"],
   transform: {
     "^.+\\.(js|mjs|jsx|ts|tsx)$": "babel-jest",
@@ -42,9 +23,50 @@ const config: Config = {
   transformIgnorePatterns: [
     `<rootDir>/node_modules/(?!(${esmOnlyPackages.join("|")}))`,
   ],
+  moduleDirectories: ["src", "node_modules"],
   moduleNameMapper: {
     "\\.(css|less)$": "<rootDir>/__mocks__/styleMock.js",
   },
+  coveragePathIgnorePatterns: [
+    "<rootDir>/node_modules",
+    "<rootDir>/src/__spec_helper__",
+    "<rootDir>/src/locales",
+    "<rootDir>/lib",
+    "<rootDir>/esm",
+  ],
+  coverageDirectory: "<rootDir>/coverage",
+  testPathIgnorePatterns: [
+    "<rootDir>/node_modules",
+    "<rootDir>/lib",
+    "<rootDir>/esm",
+  ],
 };
 
-export default config;
+const clientConfig: Config = {
+  displayName: "Client",
+  testMatch: ["**/!(*.server).+(spec|test).[jt]s?(x)"],
+  testEnvironment: "jsdom",
+  setupFilesAfterEnv: [
+    "<rootDir>/src/__spec_helper__/__internal__/index.ts",
+    "jest-canvas-mock",
+  ],
+  ...baseProjectConfig,
+};
+
+const serverConfig: Config = {
+  displayName: { name: "Server", color: "blue" },
+  testMatch: ["**/*.server.(spec|test).[jt]s?(x)"],
+  testEnvironment: "node",
+  setupFilesAfterEnv: ["<rootDir>/src/__spec_helper__/__internal__/index.ts"],
+  ...baseProjectConfig,
+};
+
+const globalConfig: Config = {
+  projects: [clientConfig, serverConfig],
+  notify: false,
+  collectCoverage: true,
+  coverageReporters: ["text-summary", "html", "json"],
+  coverageThreshold: isCI ? undefined : coverageThresholds,
+};
+
+export default globalConfig;