+++ src/components/action-popover/components.test-pw.tsx
@@ -1,4 +1,3 @@
-/* eslint-disable react/prop-types */
 import React, { useState } from "react";
 import Box from "../box";
 import Link from "../link";