+++ src/components/icon-button/icon-button.pw.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { test, expect } from "@playwright/experimental-ct-react";
+import { test, expect } from "../../../playwright/helpers/base-test";
 import IconButtonComponent from "./component.test-pw";
 import { button as iconButton } from "../../../playwright/components/index";
 import { CHARACTERS } from "../../../playwright/support/constants";