+++ src/components/progress-tracker/progress-tracker.pw.tsx
@@ -1,7 +1,7 @@
-import { expect, test } from "@playwright/experimental-ct-react";
 import React from "react";
 
 import { Locator } from "@playwright/test";
+import { test, expect } from "../../../playwright/helpers/base-test";
 import {
   progressTrackerComponent,
   progressTrackerCustomValuePreposition,